Replace IntFlags with Uint64Flags in TCP/UDP

+++ beacon-chain/p2p/config.go
@@ -21,8 +21,8 @@ type Config struct {
 	PrivateKey          string
 	DataDir             string
 	MetaDataDir         string
-	TCPPort             uint
-	UDPPort             uint
+	TCPPort             uint64
+	UDPPort             uint64
 	MaxPeers            uint64
 	AllowListCIDR       string
 	DenyListCIDR        []string
