+++ cmd/flags.go
@@ -108,13 +108,13 @@ var (
 		Value: "",
 	}
 	// P2PUDPPort defines the port to be used by discv5.
-	P2PUDPPort = &cli.IntFlag{
+	P2PUDPPort = &cli.Uint64Flag{
 		Name:  "p2p-udp-port",
 		Usage: "The port used by discv5.",
 		Value: 12000,
 	}
 	// P2PTCPPort defines the port to be used by libp2p.
-	P2PTCPPort = &cli.IntFlag{
+	P2PTCPPort = &cli.Uint64Flag{
 		Name:  "p2p-tcp-port",
 		Usage: "The port used by libp2p.",
 		Value: 13000,
